payment method curd + add rule validator unique deleted_at null

+++ app/Http/Controllers/PaymentMethodController.php
@@ -2,15 +2,14 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Resources\OrderStatusCollection;
-use App\Http\Resources\OrderStatusResource;
-use App\Http\Validators\OrderStatus\OrderStatusUpsertValidator;
-use App\Models\OrderStatus;
+use App\Http\Resources\PaymentMethodCollection;
+use App\Http\Resources\PaymentMethodResource;
+use App\Http\Validators\PaymentMethod\PaymentMethodUpsertValidator;
+use App\Models\PaymentMethod;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
-use Symfony\Component\HttpKernel\Exception\HttpException;
 
-class OrderStatusController extends Controller
+class PaymentMethodController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -22,15 +21,18 @@ class OrderStatusController extends Controller
         $input = $request->all();
         $input['limit'] = $request->limit;
         try{
-            $data = OrderStatus::where('is_active', !empty($input['is_active']) ? $input['is_active'] : 1)->where(function($query) use($input) {
+            $data = PaymentMethod::where('is_active', !empty($input['is_active']) ? $input['is_active'] : 1)->where(function($query) use($input) {
                 if(!empty($input['name'])){
                     $query->where('name', 'like', '%'.$input['name'].'%');
+                }
+                if(!empty($input['code'])){
+                    $query->where('code', $input['code']);
                 }
                 if(!empty($input['is_active'])){
                     $query->where('is_active', $input['is_active']);
                 }
             })->orderBy('created_at', 'desc')->paginate(!empty($input['limit']) ? $input['limit'] : 10);
-            $resource = new OrderStatusCollection($data);
+            $resource = new PaymentMethodCollection($data);
         }
         catch(HttpException $e){
             return response()->json([
@@ -51,24 +53,16 @@ class OrderStatusController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request, OrderStatusUpsertValidator $validator)
+    public function store(Request $request, PaymentMethodUpsertValidator $validator)
     {
         $input = $request->all();
         $validator->validate($input);
         try{
-            $check = OrderStatus::where('code', $request->code)->whereNull('deleted_at')->first();
-            if(!empty($check)){
-                return response()->json([
-                    'status' => 'error',
-                    'message' => 'Trạng thái đã tồn tại !',
-                ], 400);
-            }
             DB::beginTransaction();
-            $orderStatus = OrderStatus::create([
-                'name' => $request->name,
-                'code' => $request->code,
-                'sort_level' => $request->sort_level ?? null,
-                'is_active' => $request->is_active ?? 1,
+            $paymentMethod = PaymentMethod::create([
+                'name' => $input['name'],
+                'code' => strtoupper($input['code']),
+                'is_active' => $input['is_active'] ?? 1,
                 'created_by' => auth('sanctum')->user()->id,
                 'updated_by' => auth('sanctum')->user()->id,
             ]);
@@ -87,7 +81,7 @@ class OrderStatusController extends Controller
         }
         return response()->json([
             'status' => 'success',
-            'message' => 'Trạng thái đơn hàng ['.$orderStatus->name.'] đã được tạo thành công !',
+            'message' => 'Trạng thái đơn hàng ['.$paymentMethod->name.'] đã được tạo thành công !',
         ]);
     }
 
@@ -100,11 +94,11 @@ class OrderStatusController extends Controller
     public function show($id)
     {
         try{
-            $data = OrderStatus::find($id);
+            $data = PaymentMethod::find($id);
             if(empty($data)){
                 return response()->json([
                     'status' => 'error',
-                    'message' => 'Không tìm thấy trạng thái đơn hàng !',
+                    'message' => 'Không tìm thấy phương thức thanh toán !',
                 ], 404);
             }
         }
@@ -120,7 +114,7 @@ class OrderStatusController extends Controller
         }
         return response()->json([
             'status' => 'success',
-            'data' => new OrderStatusResource($data),
+            'data' => new PaymentMethodResource($data),
         ]);
     }
 
@@ -131,32 +125,25 @@ class OrderStatusController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id, OrderStatusUpsertValidator $validator)
+    public function update(Request $request, $id, PaymentMethodUpsertValidator $validator)
     {
         $input = $request->all();
         $validator->validate($input);
         try {
-            $check = OrderStatus::where('code', $request->code)->whereNull('deleted_at')->first();
-            if(empty($check)){
-                return response()->json([
-                    'status' => 'error',
-                    'message' => 'Trạng thái đã tồn tại !',
-                ], 400);
-            }
             DB::beginTransaction();
-            $orderStatus = OrderStatus::find($id);
-            if(empty($orderStatus)) {
+            $paymentMethod = PaymentMethod::find($id);
+            if(empty($paymentMethod)){
                 return response()->json([
                     'status' => 'error',
-                    'message' => 'Trạng thái đơn hàng không tồn tại !',
+                    'message' => 'Phương thức thanh toán không tồn tại !',
                 ], 404);
             }
-            $orderStatus->name = $request->name ?? $orderStatus->name;
-            $orderStatus->code = $request->code ?? $orderStatus->code;
-            $orderStatus->sort_level = $request->sort_level ?? $orderStatus->sort_level;
-            $orderStatus->is_active = $request->is_active ?? $orderStatus->is_active;
-            $orderStatus->updated_by = auth('sanctum')->user()->id;
-            $orderStatus->save();
+            $paymentMethod->name = $request->name ?? $paymentMethod->name;
+            $paymentMethod->code = strtoupper($request->code) ?? $paymentMethod->code;
+            $paymentMethod->is_active = $request->is_active ?? $paymentMethod->is_active;
+            $paymentMethod->updated_by = auth('sanctum')->user()->id;
+            $paymentMethod->save();
+
             DB::commit();
         }
         catch(HttpException $e){
@@ -172,7 +159,7 @@ class OrderStatusController extends Controller
         }
         return response()->json([
             'status' => 'success',
-            'message' => 'Trạng thái đơn hàng ['.$orderStatus->name.'] đã được cập nhật !',
+            'message' => 'Phương thức thanh toán ['.$paymentMethod->name.'] đã được cập nhật !',
         ]);
     }
 
@@ -182,22 +169,25 @@ class OrderStatusController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy($id, Request $request)
     {
         try{
             DB::beginTransaction();
-            $data = OrderStatus::find($id);
-            if(empty($data)){
-                return response()->json([
-                    'status' => 'error',
-                    'message' => 'Trạng thái đơn hàng không tồn tại !',
-                ], 404);
+            if(!is_array($id)){
+                $data = PaymentMethod::find($id);
+                if(empty($data)){
+                    return response()->json([
+                        'status' => 'error',
+                        'message' => 'Không tìm thấy người dùng !',
+                    ], 404);
+                }
+                $data->is_delete = 1;
+                $data->deleted_by = auth('sanctum')->user()->id;
+                $data->save();
+
+                $data->delete();
             }
-            $data->update([
-                'is_delete' => 1,
-                'deleted_by' => auth('sanctum')->user()->id
-            ]);
-            $data->delete();
+
             DB::commit();
         }
         catch(HttpException $e){
